+++ nebulosa-astap/src/main/kotlin/nebulosa/astap/star/detection/Star.kt
@@ -5,7 +5,7 @@ import nebulosa.star.detection.ImageStar
 data class Star(
     override val x: Int = 0,
     override val y: Int = 0,
-    override val hfd: Float = 0f,
-    override val snr: Float = 0f,
-    override val flux: Float = 0f,
+    override val hfd: Double = 0.0,
+    override val snr: Double = 0.0,
+    override val flux: Double = 0.0,
 ) : ImageStar
