+++ nebulosa-guiding-internal/src/main/kotlin/nebulosa/guiding/internal/MassChecker.kt
@@ -6,14 +6,14 @@ internal class MassChecker {
 
     private data class Entry(
         val time: Long,
-        val mass: Float,
+        val mass: Double,
     )
 
     data class CheckedMass(
-        @JvmField val limit0: Float = 0f,
-        @JvmField val limit1: Float = 0f,
-        @JvmField val limit2: Float = 0f,
-        @JvmField val limit3: Float = 0f,
+        @JvmField val limit0: Double = 0.0,
+        @JvmField val limit1: Double = 0.0,
+        @JvmField val limit2: Double = 0.0,
+        @JvmField val limit3: Double = 0.0,
         @JvmField val reject: Boolean = true,
     ) {
 
@@ -24,8 +24,8 @@ internal class MassChecker {
     }
 
     private val data = LinkedList<Entry>()
-    private var highMass = 0f
-    private var lowMass = Float.MAX_VALUE
+    private var highMass = 0.0
+    private var lowMass = Double.MAX_VALUE
     private var exposure = 0L
     private var timeWindow = DEFAULT_TIME_WINDOW * 2L
 
@@ -41,7 +41,7 @@ internal class MassChecker {
         }
     }
 
-    fun add(mass: Float) {
+    fun add(mass: Double) {
         val now = System.currentTimeMillis()
         val oldest = now - timeWindow
 
@@ -56,7 +56,7 @@ internal class MassChecker {
         else data.add(-index - 1, entry)
     }
 
-    fun checkMass(mass: Float, threshold: Float = 0.5f): CheckedMass {
+    fun checkMass(mass: Double, threshold: Double = 0.5): CheckedMass {
         if (data.size < 5) return CheckedMass.EMPTY
 
         val median = data[data.size / 2]
@@ -67,15 +67,15 @@ internal class MassChecker {
         // Let the low water mark drift to follow the median
         // so that it moves back up after a period of intermittent
         // clouds has brought it down.
-        lowMass += 0.05f * (median.mass - lowMass)
+        lowMass += 0.05 * (median.mass - lowMass)
 
-        val limit0 = lowMass * (1f - threshold)
+        val limit0 = lowMass * (1.0 - threshold)
         val limit1 = median.mass
-        val limit2 = highMass * (1f + threshold)
+        val limit2 = highMass * (1.0 + threshold)
         // When mass is depressed by sky conditions, we still want to trigger a rejection when
         // there is a large spike in mass, even if it is still below the high water mark-based
         // threhold.
-        val limit3 = median.mass * (1f + 2f * threshold)
+        val limit3 = median.mass * (1.0 + 2.0 * threshold)
 
         val reject = mass < limit0 || mass > limit2 || mass > limit3
 
@@ -84,8 +84,8 @@ internal class MassChecker {
 
     fun reset() {
         data.clear()
-        highMass = 0f
-        lowMass = Float.MAX_VALUE
+        highMass = 0.0
+        lowMass = Double.MAX_VALUE
     }
 
     companion object {
