+++ nebulosa-imaging/hfd/src/main/kotlin/nebulosa/imaging/hfd/Star.kt
@@ -4,7 +4,7 @@ import nebulosa.star.detection.ImageStar
 
 data class Star(
     override val x: Int, override val y: Int,
-    override val hfd: Float = 0f,
-    override val snr: Float = 0f,
-    override val flux: Float = 0f,
+    override val hfd: Double = 0.0,
+    override val snr: Double = 0.0,
+    override val flux: Double = 0.0,
 ) : ImageStar
