+++ nebulosa-watney/src/main/kotlin/nebulosa/watney/plate/solving/quad/ImageStarQuad.kt
@@ -4,18 +4,18 @@ import nebulosa.star.detection.ImageStar
 
 @Suppress("ArrayInDataClass")
 data class ImageStarQuad(
-    override val ratios: FloatArray,
-    override val largestDistance: Float,
-    override val midPointX: Float,
-    override val midPointY: Float,
+    override val ratios: DoubleArray,
+    override val largestDistance: Double,
+    override val midPointX: Double,
+    override val midPointY: Double,
     override val stars: List<ImageStar> = emptyList(),
 ) : StarQuad {
 
-    constructor(ratios: FloatArray, largestDistance: Float, stars: List<ImageStar> = emptyList())
+    constructor(ratios: DoubleArray, largestDistance: Double, stars: List<ImageStar> = emptyList())
             : this(
         ratios, largestDistance,
-        (stars[0].x + stars[1].x + stars[2].x + stars[3].x) / 4f,
-        (stars[0].y + stars[1].y + stars[2].y + stars[3].y) / 4f,
+        (stars[0].x + stars[1].x + stars[2].x + stars[3].x) / 4.0,
+        (stars[0].y + stars[1].y + stars[2].y + stars[3].y) / 4.0,
         stars
     )
