+++ nebulosa-watney/src/main/kotlin/nebulosa/watney/plate/solving/quad/QuadDatabaseCellFile.kt
@@ -43,7 +43,7 @@ internal data class QuadDatabaseCellFile(@JvmField val descriptor: QuadDatabaseC
         }
 
         // Pre-allocate, so that we don't need to allocate later down the road.
-        val quadDataArray = FloatArray(8)
+        val quadDataArray = DoubleArray(8)
 
         repeat(subCellsInRangeLen) {
             val subCellIdx = subCellsInRangeIndexesArr[it]
@@ -81,7 +81,7 @@ internal data class QuadDatabaseCellFile(@JvmField val descriptor: QuadDatabaseC
                         matchingQuads.add(quad)
 
                         val distanceTo =
-                            SphericalCoordinate.angularDistance(quad.midPointX.toDouble(), quad.midPointY.toDouble(), centerRA, centerDEC)
+                            SphericalCoordinate.angularDistance(quad.midPointX, quad.midPointY, centerRA, centerDEC)
                         if (distanceTo < angularDistance) matchingQuadsWithinRange.add(quad)
                     }
                 }
@@ -96,11 +96,11 @@ internal data class QuadDatabaseCellFile(@JvmField val descriptor: QuadDatabaseC
         // RATIOS, LARGEST_DIST, COORDS
         const val DATA_LENGTH = 6L + 4 + 4 * 2
 
-        private const val TEN_BITS = 1f / 1023
-        private const val NINE_BITS = 1f / 511
+        private const val TEN_BITS = 1.0 / 1023
+        private const val NINE_BITS = 1.0 / 511
 
         @JvmStatic
-        private fun bytesToQuadNew(source: Source, tentativeMatches: List<ImageStarQuad>?, quadDataArray: FloatArray): StarQuad? {
+        private fun bytesToQuadNew(source: Source, tentativeMatches: List<ImageStarQuad>?, quadDataArray: DoubleArray): StarQuad? {
             val buffer = Buffer()
             require(source.read(buffer, DATA_LENGTH) == DATA_LENGTH) { "unexpected end of file" }
             val ratios = buffer.readByteArray(6L)
@@ -115,9 +115,9 @@ internal data class QuadDatabaseCellFile(@JvmField val descriptor: QuadDatabaseC
             var starQuad: StarQuad? = null
 
             if (tentativeMatches.isNullOrEmpty()) {
-                quadDataArray[5] = buffer.readFloat() // LARGEST_DIST
-                quadDataArray[6] = buffer.readFloat() // RA
-                quadDataArray[7] = buffer.readFloat() // DEC
+                quadDataArray[5] = buffer.readFloat().toDouble() // LARGEST_DIST
+                quadDataArray[6] = buffer.readFloat().toDouble() // RA
+                quadDataArray[7] = buffer.readFloat().toDouble() // DEC
                 starQuad = ImageStarQuad(quadDataArray.sliceArray(0..4), quadDataArray[5], quadDataArray[6], quadDataArray[7])
             } else {
                 for ((r) in tentativeMatches) {
@@ -127,9 +127,9 @@ internal data class QuadDatabaseCellFile(@JvmField val descriptor: QuadDatabaseC
                         && abs(r[3] / quadDataArray[3] - 1.0) <= 0.011
                         && abs(r[4] / quadDataArray[4] - 1.0) <= 0.011
                     ) {
-                        quadDataArray[5] = buffer.readFloat() // LARGEST_DIST
-                        quadDataArray[6] = buffer.readFloat() // RA
-                        quadDataArray[7] = buffer.readFloat() // DEC
+                        quadDataArray[5] = buffer.readFloat().toDouble() // LARGEST_DIST
+                        quadDataArray[6] = buffer.readFloat().toDouble() // RA
+                        quadDataArray[7] = buffer.readFloat().toDouble() // DEC
                         starQuad = ImageStarQuad(quadDataArray.sliceArray(0..4), quadDataArray[5], quadDataArray[6], quadDataArray[7])
                         break
                     }
