+++ nebulosa-watney/src/main/kotlin/nebulosa/watney/star/detection/Star.kt
@@ -5,7 +5,7 @@ import nebulosa.star.detection.ImageStar
 data class Star(
     override val x: Int = 0, override val y: Int = 0,
     val size: Double = 0.0,
-    override var hfd: Float = 0f,
-    override var snr: Float = 0f,
-    override var flux: Float = 0f,
+    override var hfd: Double = 0.0,
+    override var snr: Double = 0.0,
+    override var flux: Double = 0.0,
 ) : ImageStar
